Data generator creates unique records!
