+++ client/src/components/app.jsx
@@ -11,7 +11,7 @@ class App extends React.Component {
   }
 
   componentDidMount() {
-    axios.get('http://3.18.221.86:3004/api/products')
+    axios.get('http://localhost:3004/products')
       .then((result) => {
         this.setState({
           similarProducts: result.data
