+++ dataGen/products.csv
@@ -1,0 +1,10 @@
+name,image_url,price,description,category_id
+"Sleek Granite Table","http://placeimg.com/640/480/fashion","432.00","Carbonite web goalkeeper gloves are ergonomically designed to give easy fit",4064
+"Handmade Wooden Salad","http://placeimg.com/640/480/city","564.00","The automobile layout consists of a front-engine design, with transaxle-type transmissions mounted at the rear of the engine and four wheel drive",5462
+"Handmade Frozen Keyboard","http://placeimg.com/640/480/fashion","928.00","Andy shoes are designed to keeping in mind durability as well as trends, the most stylish range of shoes & sandals",8484
+"Practical Frozen Tuna","http://placeimg.com/640/480/food","934.00","The Apollotech B340 is an affordable wireless mouse with reliable connectivity, 12 months battery life and modern design",6146
+"Licensed Wooden Salad","http://placeimg.com/640/480/sports","625.00","The Nagasaki Lander is the trademarked name of several series of Nagasaki sport bikes, that started with the 1984 ABC800J",3461
+"Generic Metal Bike","http://placeimg.com/640/480/cats","262.00","The Football Is Good For Training And Recreational Purposes",6131
+"Tasty Fresh Soap","http://placeimg.com/640/480/technics","673.00","New ABC 13 9370, 13.3, 5th Gen CoreA5-8250U, 8GB RAM, 256GB SSD, power UHD Graphics, OS 10 Home, OS Office A & J 2016",7701
+"Sleek Cotton Mouse","http://placeimg.com/640/480/business","725.00","Andy shoes are designed to keeping in mind durability as well as trends, the most stylish range of shoes & sandals",997
+"Refined Rubber Table","http://placeimg.com/640/480/city","693.00","The beautiful range of Apple Naturalé that has an exciting mix of natural ingredients. With the Goodness of 100% Natural Ingredients",8976
