--- databasePg/products.csv
@@ -1,10 +1,0 @@
-name,image_url,price,description,category_id"Gorgeous Cotton Shoes","http://placeimg.com/640/480/cats","43.00","The Football Is Good For Training And Recreational Purposes",2194
-"Gorgeous Cotton Shoes","http://placeimg.com/640/480/cats","43.00","The Football Is Good For Training And Recreational Purposes",2194
-"Gorgeous Cotton Shoes","http://placeimg.com/640/480/cats","43.00","The Football Is Good For Training And Recreational Purposes",2194
-"Gorgeous Cotton Shoes","http://placeimg.com/640/480/cats","43.00","The Football Is Good For Training And Recreational Purposes",2194
-"Gorgeous Cotton Shoes","http://placeimg.com/640/480/cats","43.00","The Football Is Good For Training And Recreational Purposes",2194
-"Gorgeous Cotton Shoes","http://placeimg.com/640/480/cats","43.00","The Football Is Good For Training And Recreational Purposes",2194
-"Gorgeous Cotton Shoes","http://placeimg.com/640/480/cats","43.00","The Football Is Good For Training And Recreational Purposes",2194
-"Gorgeous Cotton Shoes","http://placeimg.com/640/480/cats","43.00","The Football Is Good For Training And Recreational Purposes",2194
-"Gorgeous Cotton Shoes","http://placeimg.com/640/480/cats","43.00","The Football Is Good For Training And Recreational Purposes",2194
-"Gorgeous Cotton Shoes","http://placeimg.com/640/480/cats","43.00","The Football Is Good For Training And Recreational Purposes",2194
