+++ package.json
@@ -15,6 +15,7 @@
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "faker": "^5.1.0",
+    "json2csv": "^5.0.6",
     "mongoose": "^5.11.12",
     "node-postgres": "^0.6.2",
     "nodemon": "^2.0.7",
@@ -35,7 +36,7 @@
   },
   "scripts": {
     "test": "jest --forceExit",
-    "seed": "node -e \"require('./database/seeder.js').seed(4)\"",
+    "seed": "node -e \"require('./databasePg/seeder.js').seed(4)\"",
     "build": "webpack",
     "watch": "webpack -w",
     "start": "nodemon node ./server/index.js"
