+++ server/index.js
@@ -11,3 +11,4 @@ app.listen(port, () => {
 })
 
 module.exports = app;
+
